refactor: rename logger directory in /utils to /log
