+++ server/src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from "./logger";
+export * from "./log";
